+++ lib/fc_native_image_resize_platform_interface.dart
@@ -31,7 +31,7 @@ abstract class FcNativeImageResizePlatform extends PlatformInterface {
       required int height,
       required bool keepAspectRatio,
       required String type,
-      double? quality}) {
+      int? quality}) {
     throw UnimplementedError('platformVersion() has not been implemented.');
   }
 }
